requirements.txt

django-ipware==4.0.2

APScheduler==3.9.0
beautifulsoup4==4.10.0
psutil==5.9.4
Django==4.0.1
django-admin-interface==0.16.3
django-background-tasks
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-filter==2.4.0
django-import-export==2.7.1
django-password-reset==2.0
django-registration==3.1.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
fcm-django==1.0.8
grpcio==1.49.0
gunicorn==19.10.0
heroku==0.1.4
inapppy==2.5.2
lxml==4.6.3
mysql-connector==2.2.9
Pillow==8.0.1
play-scraper==0.6.0
psycopg2-binary==2.8.6
py==1.10.0
pybase64==1.1.4
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2020.5
requests==2.25.1
requests-futures==1.0.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1
urlparser==0.1.2
validate-email==1.3
whitenoise==5.2.0
xlrd==2.0.1
XlsxWriter==1.3.6
xlwt==1.3.0